requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
PyJWT>=2.9
minio>=7.2
ulid-py>=1.1
python-dotenv>=1.0
